+++ app/src/main/java/com/example/plantee/ui/nav/Screens.kt
@@ -10,16 +10,16 @@ sealed interface Screen : NavKey {
     @Serializable
     data object DiagnosePlant : Screen
     @Serializable
-    data class DiagnosisDetails(val id: Int) : Screen
+    data class DiagnosisDetails(val id: Long) : Screen
     @Serializable
-    data class DiagnosisResults(val id: Int) : Screen
+    data class DiagnosisResults(val id: Long) : Screen
 
     @Serializable
     data object PlantAdd : Screen
     @Serializable
-    data class PlantDetails(val id: Int) : Screen
+    data class PlantDetails(val id: Long) : Screen
     @Serializable
-    data class PlantEdit(val id: Int): Screen
+    data class PlantEdit(val id: Long): Screen
     @Serializable
     data object Plants : Screen
 
@@ -27,9 +27,9 @@ sealed interface Screen : NavKey {
     @Serializable
     data object RoutineAdd : Screen
     @Serializable
-    data class RoutineDetails(val id: Int) : Screen
+    data class RoutineDetails(val id: Long) : Screen
     @Serializable
-    data class RoutineEdit(val id: Int) : Screen
+    data class RoutineEdit(val id: Long) : Screen
     @Serializable
     data object Routines : Screen
